Create courses info component
